Update to QT6
Replace removed obsolete functions with their recomendend replacements.
Some of them are not available in QT4, so it is no longer possible
to build against it. This should not cause any troubles, as QT4 is
unsupported since 2015.

Tested only on Debian 12 with with Qt 6.4.2, so project file is left
untouched, until more platforms are verified to build and run.